requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
line-bot-sdk==3.14.3
google-genai==1.14.0
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
python-dotenv==1.1.0
aiohttp==3.11.18
apscheduler==3.11.0
Pillow==11.2.1
